+++ src/dockerCommands.ts
@@ -23,6 +23,10 @@ export class DockerCommands {
 	public analyzeDiagnostics(diagnostics: Diagnostic[], textDocumentURI: string, range: Range): Command[] {
 		let commands = [];
 		for (let i = 0; i < diagnostics.length; i++) {
+			// Diagnostic's code is (number | string), convert it if necessary
+			if (typeof diagnostics[i].code === "string") {
+				diagnostics[i].code = parseInt(diagnostics[i].code as string);
+			}
 			switch (diagnostics[i].code) {
 				case ValidationCode.CASING_INSTRUCTION:
 					commands.push({